Codechange: split off the settings table and all the callbacks from the main settings handling logic
