+++ src/CMakeLists.txt
@@ -338,6 +338,8 @@ add_files(
     settings_gui.cpp
     settings_gui.h
     settings_internal.h
+    settings_table.h
+    settings_table.cpp
     settings_type.h
     ship.h
     ship_cmd.cpp
