+++ src/table/settings/company_settings.ini
@@ -12,7 +12,7 @@ static void UpdateAllServiceInterval(int32 new_value);
 static bool CanUpdateServiceInterval(VehicleType type, int32 &new_value);
 static void UpdateServiceInterval(VehicleType type, int32 new_value);
 
-static const SettingVariant _company_settings[] = {
+static const SettingVariant _company_settings_table[] = {
 [post-amble]
 };
 [templates]
