+++ src/table/settings/currency_settings.ini
@@ -7,7 +7,7 @@
 ; Settings for the in-game custom currency.
 
 [pre-amble]
-static const SettingVariant _currency_settings[] = {
+static const SettingVariant _currency_settings_table[] = {
 [post-amble]
 };
 [templates]
