+++ src/table/settings/gameopt_settings.ini
@@ -14,9 +14,8 @@
 ; be saved in their new place.
 
 [pre-amble]
-static const uint GAME_DIFFICULTY_NUM = 18;
-static const std::array<std::string, GAME_DIFFICULTY_NUM> _old_diff_settings{"max_no_competitors", "competitor_start_time", "number_towns", "industry_density", "max_loan", "initial_interest", "vehicle_costs", "competitor_speed", "competitor_intelligence", "vehicle_breakdowns", "subsidy_multiplier", "construction_cost", "terrain_type", "quantity_sea_lakes", "economy", "line_reverse_mode", "disasters", "town_council_tolerance"};
-static uint16 _old_diff_custom[GAME_DIFFICULTY_NUM];
+const std::array<std::string, GAME_DIFFICULTY_NUM> _old_diff_settings{"max_no_competitors", "competitor_start_time", "number_towns", "industry_density", "max_loan", "initial_interest", "vehicle_costs", "competitor_speed", "competitor_intelligence", "vehicle_breakdowns", "subsidy_multiplier", "construction_cost", "terrain_type", "quantity_sea_lakes", "economy", "line_reverse_mode", "disasters", "town_council_tolerance"};
+uint16 _old_diff_custom[GAME_DIFFICULTY_NUM];
 uint8 _old_diff_level;                                 ///< Old difficulty level from old savegames
 uint8 _old_units;                                      ///< Old units from old savegames
 
@@ -33,7 +32,7 @@ static std::initializer_list<const char*> _osk_activation{"disabled", "double", 
 static std::initializer_list<const char*> _settings_profiles{"easy", "medium", "hard"};
 static std::initializer_list<const char*> _news_display{ "off", "summarized", "full"};
 
-static const SettingVariant _gameopt_settings[] = {
+static const SettingVariant _gameopt_settings_table[] = {
 /* In version 4 a new difficulty setting has been added to the difficulty settings,
  * town attitude towards demolishing. Needs special handling because some dimwit thought
  * it funny to have the GameDifficulty struct be an array while it is a struct of
