+++ src/table/settings/misc_settings.ini
@@ -20,7 +20,7 @@ extern bool _allow_hidpi_window;
 #define WITHOUT_COCOA
 #endif
 
-static const SettingVariant _misc_settings[] = {
+static const SettingVariant _misc_settings_table[] = {
 [post-amble]
 };
 [templates]
