+++ src/table/settings/network_private_settings.ini
@@ -7,7 +7,7 @@
 ; Network settings as stored in the private configuration file ("private.cfg").
 
 [pre-amble]
-static const SettingVariant _network_private_settings[] = {
+static const SettingVariant _network_private_settings_table[] = {
 [post-amble]
 };
 [templates]
