+++ src/table/settings/network_secrets_settings.ini
@@ -9,7 +9,7 @@
 [pre-amble]
 static bool ReplaceAsteriskWithEmptyPassword(std::string &newval);
 
-static const SettingVariant _network_secrets_settings[] = {
+static const SettingVariant _network_secrets_settings_table[] = {
 [post-amble]
 };
 [templates]
