+++ src/table/settings/network_settings.ini
@@ -9,7 +9,7 @@
 [pre-amble]
 static void UpdateClientConfigValues();
 
-static const SettingVariant _network_settings[] = {
+static const SettingVariant _network_settings_table[] = {
 [post-amble]
 };
 [templates]
