+++ src/table/settings/settings.ini
@@ -46,7 +46,7 @@ static void InvalidateShipPathCache(int32 new_value);
  * assigns its own value. If the setting was company-based, that would mean that
  * vehicles could decide on different moments that they are heading back to a
  * service depot, causing desyncs on a massive scale. */
-static const SettingVariant _settings[] = {
+static const SettingVariant _settings_table[] = {
 [post-amble]
 };
 [templates]
