+++ src/table/settings/win32_settings.ini
@@ -12,7 +12,7 @@
 #if defined(_WIN32) && !defined(DEDICATED)
 extern bool _window_maximize;
 
-static const SettingVariant _win32_settings[] = {
+static const SettingVariant _win32_settings_table[] = {
 [post-amble]
 };
 #endif /* _WIN32 */
