+++ src/table/settings/window_settings.ini
@@ -9,7 +9,7 @@
 
 [pre-amble]
 
-static const SettingVariant _window_settings[] = {
+static const SettingVariant _window_settings_table[] = {
 [post-amble]
 };
 [templates]
